Export DeduplicateFindings from pkg/models
Move finding dedup (group by fingerprint, consolidate locations) and the
location formatter out of pkg/reporter into pkg/models as exported
DeduplicateFindings and FindingLocation. The reporter now delegates to
them, so library consumers get the same finding-identity semantics without
importing the reporter (and its table-rendering deps). Dedup test moves to
pkg/models alongside the logic.
